requirements.txt
dashscope
streamlit
python-dotenv
pypdf
openai